requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
pymongo==3.12.1
pythondi==1.2.2
python-dotenv==0.19.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
redis==4.0.2